// old-v7/packages/core/src/lib/types.ts
export type NgtBooleanInput = string | boolean | null | undefined;
export type NgtNumberInput = string | number | null | undefined;
export type NgtObservableInput<T> = T | ObservableInput<T>;
// ...
export type NgtUnknownRecord = Record<string, unknown>;
// ...
export type NgtCamera = THREE.OrthographicCamera | THREE.PerspectiveCamera;
// ...
export type NgtThreeEvent<TEvent> = NgtIntersectionEvent<TEvent>;

export interface NgtEventHandlers {
  click?: (event: NgtThreeEvent<MouseEvent>) => void;
  contextmenu?: (event: NgtThreeEvent<MouseEvent>) => void;
  dblclick?: (event: NgtThreeEvent<MouseEvent>) => void;
  pointerup?: (event: NgtThreeEvent<PointerEvent>) => void;
  pointerdown?: (event: NgtThreeEvent<PointerEvent>) => void;
  pointerover?: (event: NgtThreeEvent<PointerEvent>) => void;
  pointerout?: (event: NgtThreeEvent<PointerEvent>) => void;
  pointerenter?: (event: NgtThreeEvent<PointerEvent>) => void;
  pointerleave?: (event: NgtThreeEvent<PointerEvent>) => void;
  pointermove?: (event: NgtThreeEvent<PointerEvent>) => void;
  pointermissed?: (event: MouseEvent) => void;
  pointercancel?: (event: NgtThreeEvent<PointerEvent>) => void;
  wheel?: (event: NgtThreeEvent<WheelEvent>) => void;
}
// ...
export interface NgtEventManager<TTarget> {
  /** Determines if the event layer is active */
  enabled: boolean;
  /** Event layer priority, higher prioritized layers come first and may stop(-propagate) lower layer  */
  priority: number;
  /** The compute function needs to set up the raycaster and an xy- pointer  */
  compute?: NgtComputeFunction;
  /** The filter can re-order or re-structure the intersections  */
  filter?: NgtFilterFunction;
  /** The target node the event layer is tied to */
  connected?: TTarget;
  /** All the pointer event handlers through which the host forwards native events */
  handlers?: NgtEvents;
  /** Allows re-connecting to another target */
  connect?: (target: TTarget) => void;
  /** Removes all existing events handlers from the target */
  disconnect?: () => void;
}

export interface NgtInternalState {
  active: boolean;
  priority: number;
  frames: number;
  lastEvent: NgtDomEvent | null;
  interaction: THREE.Object3D[];
  hovered: Map<string, NgtThreeEvent<NgtDomEvent>>;
  animations: Map<string, NgtBeforeRenderRecord>;
  subscribers: NgtBeforeRenderRecord[];
  capturedMap: Map<number, Map<THREE.Object3D, NgtPointerCaptureTarget>>;
  initialClick: [x: number, y: number];
  initialHits: THREE.Object3D[];
}

export type NgtDpr = number | [min: number, max: number];
export interface NgtSize {
  width: number;
  height: number;
  top: number;
  left: number;
  updateStyle?: boolean;
}

export interface NgtViewport extends NgtSize {
  /** The initial pixel ratio */
  initialDpr: number;
  /** Current pixel ratio */
  dpr: number;
  /** size.width / viewport.width */
  factor: number;
  /** Camera distance */
  distance: number;
  /** Camera aspect ratio: width / height */
  aspect: number;
}

export interface NgtPerformanceOptions {
  current?: number;
  min?: number;
  max?: number;
  debounce?: number;
}

export interface NgtPerformance extends Required<NgtPerformanceOptions> {
  regress: () => void;
}

export interface NgtState {
  ready: boolean;
  /** Render loop flags */
  frameloop: 'always' | 'demand' | 'never';
  /* WebGLRenderer instance */
  gl: THREE.WebGLRenderer;
  /** Default camera */
  camera: NgtCamera & { manual?: boolean };
  cameraRef: NgtRef<NgtCamera & { manual?: boolean }>;
  /** Default scene */
  scene: THREE.Scene;
  sceneRef: NgtRef<THREE.Scene>;
  /** Normalized event coordinates */
  pointer: THREE.Vector2;
  /** Default raycaster */
  raycaster: THREE.Raycaster;
  /** Default clock */
  clock: THREE.Clock;
  /** Reactive pixel-size of the canvas */
  size: NgtSize;
  /** Reactive size of the viewport in threejs units */
  viewport: NgtViewport & {
    getCurrentViewport: (
      camera?: NgtCamera,
      target?: THREE.Vector3 | Parameters<THREE.Vector3['set']>,
      size?: NgtSize
    ) => Omit<NgtViewport, 'dpr' | 'initialDpr'>;
  };
  /** Adaptive performance interface */
  performance: NgtPerformance;
  /* Whether to enable r139's THREE.ColorManagement.legacyMode */
  legacy: boolean;
  /** Shortcut to gl.outputEncoding = LinearEncoding */
  linear: boolean;
  /** Shortcut to gl.toneMapping = NoTonemapping */
  flat: boolean;
  /** Event layer interface, contains the event handler and the node they're connected to */
  events: NgtEventManager<any>;
  /** Currently used controls */
  controls: THREE.EventDispatcher | null;
  /** XR interface */
  xr: { connect: () => void; disconnect: () => void };
  /** When the canvas was clicked but nothing was hit */
  onPointerMissed?: (event: MouseEvent) => void;
  /* internals */
  internal: NgtInternalState;
  /** If this state model is layerd (via createPortal) then this contains the previous layer */
  previousRoot?: NgtStateGetter;
  /** Flags the canvas for render, but doesn't render in itself */
  invalidate: (frames?: number) => void;
  /** Advance (render) one step */
  advance: (timestamp: number, runGlobalCallbacks?: boolean) => void;
  setPerformanceCurrent: (current: number) => void;
  setEvents: (events: Partial<NgtEventManager<any>>) => void;
  setFrameloop: (frameloop: NgtState['frameloop']) => void;
  setSize: (width: number, height: number, top?: number, left?: number, updateStyle?: boolean) => void;
  setDpr: (dpr: NgtDpr) => void;
  setCamera: (camera: NgtCamera) => void;
  addInteraction(interaction: THREE.Object3D): void;
  removeInteraction(uuid: string): void;
}

export type NgtStateGetter = () => NgtState;
// ...
export type NgtAttachFunction = (parent: NgtRef, child: NgtRef) => void | (() => void);

export interface NgtInstanceLocalState {
  /** the state getter of the canvas that the instance is being rendered to */
  stateGetter: NgtStateGetter;
  rootGetter: NgtStateGetter;
  // non-object 3d
  instancesRefs: NgtRef<NgtRef[]>;
  // object-3d
  objectsRefs: NgtRef<NgtRef[]>;
  parentRef: NgtRef<NgtInstanceNode> | null;
  isPrimitive?: boolean;
  eventCount: number;
  handlers: Partial<NgtEventHandlers>;
  args?: unknown[];
  attach?: string[] | (() => void);
  attachValue?: unknown;
  memoized?: NgtUnknownRecord;
}

export type NgtInstanceNode<TNode = any> = TNode & {
  __ngt__: NgtInstanceLocalState;
  [key: string]: any;
};

export type NgtPrepareInstanceFn<TInstance> = (
  instance: TInstance,
  prepareOptions?: {
    uuid?: string;
    parentStateGetter?: NgtStateGetter;
    rootStateGetter?: NgtStateGetter;
  }
) => NgtInstanceNode<TInstance>;